% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_coverage_normal}
\alias{calc_coverage_normal}
\title{Calculate true coverage for Normal distribution}
\usage{
calc_coverage_normal(x = NULL, k = NULL, z_upper = NULL, z_lower = NULL)
}
\arguments{
\item{x}{Data from which to calculate the sample mean and standard deviation.}

\item{k}{Number of standard deviations from the mean to be used to calculate interval.}

\item{z_upper}{Z-score for upper limit of interval.}

\item{z_lower}{Z-score for lower limit of interval.}
}
\value{
Proportion of standard Normal distribution enclosed within calculated or provided sample
interval
}
\description{
Calculate true coverage for Normal distribution
}
\details{
User must provide either \code{x} and \code{k} OR \code{z_upper} and \code{z_lower}.
}
