% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-color_palette.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Color palette for package graphics}
\format{
A list with 7 elements
}
\usage{
color_palette
}
\description{
This data provides a set of pre-specified color variables for generating plots and figures
}
\keyword{datasets}
