% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_reliability.R
\name{estimate_reliability}
\alias{estimate_reliability}
\title{Estimate coverage reliability}
\usage{
estimate_reliability(
  n,
  k,
  prox_lo,
  prox_hi,
  n_sim = 1e+05,
  seed = NA,
  save_data = FALSE
)
}
\arguments{
\item{n}{Sample size.}

\item{k}{Number of standard deviations from the mean to be used to calculate interval.}

\item{prox_lo}{Lower proximity limit expressed as a proportion on the closed interval between 0
and 1.}

\item{prox_hi}{Upper proximity limit expressed as a proportion on the closed interval between 0
and 1.}

\item{n_sim}{Number of simulation iterations to generate coverage sampling distribution.}

\item{seed}{Randomization seed. Defaults to \code{NA} so that no seed is used.}

\item{save_data}{If \code{TRUE}, return vector of coverage calculations for simulated samples.}
}
\value{
Estimated reliability (\code{reliability_hat}) conditional on proximity limits. If \code{save_data}
is \code{TRUE}, return a list with \code{reliability_hat} and vector of simulated sample coverage values.
}
\description{
Given proximity range and sample size, estimate coverage reliability using Monte
Carlo simulation.
}
\examples{
estimate_reliability(
  n = 20,
  k = 2,
  prox_lo = 0.93,
  prox_hi = 0.97
)
}
