% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_across}
\alias{eval_across}
\title{Evaluate function across list elements}
\usage{
eval_across(list_object, fun)
}
\arguments{
\item{list_object}{List of objects.}

\item{fun}{String representing function that returns \code{TRUE} or \code{FALSE} to be evaluated across
\code{list_object} elements.}
}
\value{
\code{TRUE} if all elements of \code{list_object} return \code{TRUE} when evaluated by \code{fun}; otherwise,
\code{FALSE}
}
\description{
Evaluate function across list elements
}
\examples{
eval_across(
  list_object = list(2, 3),
  fun = "is.numeric"
)
}
