% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_n_ksigma.R
\name{find_n_ksigma}
\alias{find_n_ksigma}
\title{Find N for k-sigma sample interval}
\usage{
find_n_ksigma(
  proximity_range,
  reliability,
  k = NULL,
  search_interval = c(3, 500),
  n_search = NA,
  n_confirm = 1e+05,
  tolerance = 0.001,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{proximity_range}{Range within which sample coverage should fall with pre-specified
\code{reliability}.}

\item{reliability}{Targeted probability that individual coverage of parametric sample interval
applied to a Normal distribution falls within \code{proximity_range}.}

\item{k}{Standard deviation multiplier for parametric sample interval applied to Normal
distribution (typically 2 or 3). If \code{NULL}, \code{k} is assigned based on the midpoint of the
\code{proximity_range}.}

\item{search_interval}{Initial interval of sample sizes for bisection algorithm. Default is
c(3, 500).}

\item{n_search}{Number of simulation iterations during initial search for root. If \code{NA}, will be
set to approximately 10\% of \code{n_confirm} or 1000, whichever is greater.}

\item{n_confirm}{Number of simulation iterations used to confirm \code{reliability}.}

\item{tolerance}{Threshold for difference between targeted and estimated \code{reliability} that
indicates sufficient proximity of the root-finding function to 0, triggering the end of the
search. Default is 0.001.}

\item{plot}{If \code{TRUE}, plot sampling distribution for true interval coverage using sample size
identified through bisection.}

\item{verbose}{If \code{TRUE}, print the search interval used for each round of the bisection
algorithm search.}
}
\value{
A list with the following elements:
\itemize{
\item The estimated sample size required to achieve target reliability.
\item If \code{plot} is \code{TRUE}, the plot object.
}
}
\description{
Find required sample size for "k-sigma sample interval" (i.e., parametric sample
interval applied to a Normal distribution) to achieve targeted reliability using bisection
algorithm
}
\examples{
find_n_ksigma(
  proximity_range = c(0.93, 0.97),
  reliability = 0.7,
  k = 2
)
}
