% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reliability.R
\name{plot_reliability}
\alias{plot_reliability}
\title{Plot sample coverage sampling distribution}
\usage{
plot_reliability(
  n,
  k,
  prox_lo,
  prox_hi,
  n_sim,
  seed,
  color_list = color_palette
)
}
\arguments{
\item{n}{Sample size.}

\item{k}{Number of standard deviations from the mean to be used to calculate interval.}

\item{prox_lo}{Lower proximity limit expressed as a proportion on the closed interval between 0
and 1.}

\item{prox_hi}{Upper proximity limit expressed as a proportion on the closed interval between 0
and 1.}

\item{n_sim}{Number of simulation iterations to generate coverage sampling distribution.}

\item{seed}{Randomization seed.}

\item{color_list}{List object of package colors for plotting.}
}
\value{
ggplot2 histogram with highlighted acceptable coverage region
}
\description{
Plot sample coverage sampling distribution
}
