% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-color_palette.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{Package Color Palette Data}
\format{
A list of named color values
}
\usage{
color_palette
}
\description{
This list provides a selection of colors to be used in plots and other graphics
}
\keyword{datasets}
